requirements.txt
asttokens==2.4.0
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.3.0
contourpy==1.1.1
cycler==0.12.0
decorator==5.1.1
executing==2.0.0
fonttools==4.43.0
idna==3.4
ipython==8.16.1
jedi==0.19.1
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.0
matplotlib-inline==0.1.6
neo4j==5.13.0
numpy==1.26.0
packaging==23.1
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==23.9.7
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
stack-data==0.6.3
threadpoolctl==3.2.0
traitlets==5.11.2
tzdata==2023.3
urllib3==2.0.6
wcwidth==0.2.8
